backend/back-end/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.1
babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.6
django-allauth==0.41.0
django-cors-headers==3.2.1
django-extensions==2.2.9
django-multiselectfield==0.1.12
django-mysql==3.5.0
django-oauth-toolkit==1.3.2
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.1.0
django-social-auth==0.7.28
djangorestframework==3.11.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
httplib2==0.17.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2==1.9.0.post1
oauthlib==3.1.0
phonenumbers==8.12.2
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python-openid==2.2.5
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
wrapt==1.12.1